+++ Sources/Entities/BaseItem.swift
@@ -13,28 +13,28 @@ public struct BaseItem: Codable, Hashable {
     public var container: String?
     public var dateLastSaved: Date?
     public var extraIDs: [String]?
-    public var height: Int32?
+    public var height: Int?
     public var isHD: Bool?
     public var isShortcut: Bool?
     /// Gets or sets the remote trailers.
     public var remoteTrailers: [MediaURL]?
     public var shortcutPath: String?
-    public var size: Int64?
+    public var size: Int?
     public var isSupportsExternalTransfer: Bool?
-    public var width: Int32?
+    public var width: Int?
 
     public init(
         container: String? = nil,
         dateLastSaved: Date? = nil,
         extraIDs: [String]? = nil,
-        height: Int32? = nil,
+        height: Int? = nil,
         isHD: Bool? = nil,
         isShortcut: Bool? = nil,
         remoteTrailers: [MediaURL]? = nil,
         shortcutPath: String? = nil,
-        size: Int64? = nil,
+        size: Int? = nil,
         isSupportsExternalTransfer: Bool? = nil,
-        width: Int32? = nil
+        width: Int? = nil
     ) {
         self.container = container
         self.dateLastSaved = dateLastSaved
@@ -54,14 +54,14 @@ public struct BaseItem: Codable, Hashable {
         self.container = try values.decodeIfPresent(String.self, forKey: "Container")
         self.dateLastSaved = try values.decodeIfPresent(Date.self, forKey: "DateLastSaved")
         self.extraIDs = try values.decodeIfPresent([String].self, forKey: "ExtraIds")
-        self.height = try values.decodeIfPresent(Int32.self, forKey: "Height")
+        self.height = try values.decodeIfPresent(Int.self, forKey: "Height")
         self.isHD = try values.decodeIfPresent(Bool.self, forKey: "IsHD")
         self.isShortcut = try values.decodeIfPresent(Bool.self, forKey: "IsShortcut")
         self.remoteTrailers = try values.decodeIfPresent([MediaURL].self, forKey: "RemoteTrailers")
         self.shortcutPath = try values.decodeIfPresent(String.self, forKey: "ShortcutPath")
-        self.size = try values.decodeIfPresent(Int64.self, forKey: "Size")
+        self.size = try values.decodeIfPresent(Int.self, forKey: "Size")
         self.isSupportsExternalTransfer = try values.decodeIfPresent(Bool.self, forKey: "SupportsExternalTransfer")
-        self.width = try values.decodeIfPresent(Int32.self, forKey: "Width")
+        self.width = try values.decodeIfPresent(Int.self, forKey: "Width")
     }
 
     public func encode(to encoder: Encoder) throws {
