+++ Sources/Entities/PingRequestDto.swift
@@ -11,15 +11,15 @@ import Foundation
 /// Class PingRequestDto.
 public struct PingRequestDto: Codable, Hashable {
     /// Gets or sets the ping time.
-    public var ping: Int64?
+    public var ping: Int?
 
-    public init(ping: Int64? = nil) {
+    public init(ping: Int? = nil) {
         self.ping = ping
     }
 
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
-        self.ping = try values.decodeIfPresent(Int64.self, forKey: "Ping")
+        self.ping = try values.decodeIfPresent(Int.self, forKey: "Ping")
     }
 
     public func encode(to encoder: Encoder) throws {
