+++ Sources/Entities/ReadyRequestDto.swift
@@ -15,11 +15,11 @@ public struct ReadyRequestDto: Codable, Hashable {
     /// Gets or sets the playlist item identifier of the playing item.
     public var playlistItemID: String?
     /// Gets or sets the position ticks.
-    public var positionTicks: Int64?
+    public var positionTicks: Int?
     /// Gets or sets when the request has been made by the client.
     public var when: Date?
 
-    public init(isPlaying: Bool? = nil, playlistItemID: String? = nil, positionTicks: Int64? = nil, when: Date? = nil) {
+    public init(isPlaying: Bool? = nil, playlistItemID: String? = nil, positionTicks: Int? = nil, when: Date? = nil) {
         self.isPlaying = isPlaying
         self.playlistItemID = playlistItemID
         self.positionTicks = positionTicks
@@ -30,7 +30,7 @@ public struct ReadyRequestDto: Codable, Hashable {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
         self.isPlaying = try values.decodeIfPresent(Bool.self, forKey: "IsPlaying")
         self.playlistItemID = try values.decodeIfPresent(String.self, forKey: "PlaylistItemId")
-        self.positionTicks = try values.decodeIfPresent(Int64.self, forKey: "PositionTicks")
+        self.positionTicks = try values.decodeIfPresent(Int.self, forKey: "PositionTicks")
         self.when = try values.decodeIfPresent(Date.self, forKey: "When")
     }
